Do not hardcode commit hash for dcat extension
Installing DCAT extension from the formula is achieved by default with a
clone of the extension's Github repository. Moreover, the revision to
use is hardcoded in the formula.

However, one may want to install a fork of the extension instead, thus
the revision to use may be different. We remove it from the formula, and
document that a specific revision can be specified using pillars (if not
specified, head of master branch will be used).